wip: added models, contract, service

+++ backend/CrispyOctoChainsaw/CrispyOctoChainsaw.DataAccess.Postgres/Repositories/CourseRepository.cs
@@ -23,7 +23,7 @@ namespace CrispyOctoChainsaw.DataAccess.Postgres.Repositories
             return _mapper.Map<CourseEntity[], Course[]>(courses);
         }
 
-        public async Task<Course?> Get(long courseId)
+        public async Task<Course?> Get(int courseId)
         {
             var course = await _context.Courses.AsNoTracking().FirstOrDefaultAsync(x => x.Id == courseId);
             if (course is null)
